simple-agent-langgraph/requirements.txt
# Core
llama-stack-client==0.3.1
python-dotenv==1.2.1

# LangGraph and LangChain
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.14
langchain-core==1.1.1
langchain-openai==1.1.0

# FastAPI
fastapi==0.115.5
uvicorn==0.34.0
pydantic==2.10.3
email-validator==2.2.0
